fix: add restrict to relation data

+++ usecase/v1/gender_usecase.go
@@ -1,6 +1,7 @@
 package v1
 
 import (
+	"errors"
 	"strings"
 
 	"github.com/google/uuid"
@@ -177,6 +178,11 @@ func (uc GenderUsecase) Delete(genderId uuid.UUID) (err error) {
 	if err := tx.Error; err != nil {
 		logruslogger.Log(logruslogger.WarnLevel, err.Error(), functioncaller.PrintFuncName(), "gorm-start-transaction")
 		return err
+	}
+	count := tx.Model(model).Association("Products").Count()
+	if count > 0 {
+		logruslogger.Log(logruslogger.WarnLevel, "data in used", functioncaller.PrintFuncName(), "data-in-used")
+		return errors.New("data in used")
 	}
 	err = repo.Delete(model)
 	if err != nil {
